+++ tests/unit/python/test_python_submitters.py
@@ -1,6 +1,7 @@
 from unittest.mock import Mock
 
 import pytest
+from dbt_common.exceptions import DbtRuntimeError
 
 from dbt.adapters.databricks.python_models.python_submissions import (
     PythonCommandSubmitter,
@@ -9,7 +10,6 @@ from dbt.adapters.databricks.python_models.python_submissions import (
     PythonNotebookSubmitter,
     PythonNotebookUploader,
     PythonNotebookWorkflowSubmitter,
-    PythonPermissionBuilder,
     PythonWorkflowConfigCompiler,
     PythonWorkflowCreator,
 )
@@ -98,12 +98,64 @@ class TestPythonNotebookSubmitter:
         return client.job_runs.submit.return_value
 
     def test_submit__golden_path(self, submitter, compiled_code, client, tracker, run_id):
+        job_config = Mock()
+        job_config.job_spec = {}
+        job_config.additional_job_config = {}
+        submitter.config_compiler.compile.return_value = job_config
+
         submitter.submit(compiled_code)
+
         tracker.insert_run_id.assert_called_once_with(run_id)
         client.job_runs.poll_for_completion.assert_called_once_with(run_id)
         tracker.remove_run_id.assert_called_once_with(run_id)
+        client.workflow_permissions.put.assert_not_called()
+
+    def test_submit__with_acls(self, submitter, compiled_code, client, tracker, run_id):
+        job_config = Mock()
+        job_config.job_spec = {
+            "access_control_list": [{"user_name": "user", "permission_level": "CAN_RUN"}]
+        }
+        job_config.additional_job_config = {}
+        submitter.config_compiler.compile.return_value = job_config
+
+        client.job_runs.get_job_id_from_run_id.return_value = "job_id"
+
+        submitter.submit(compiled_code)
+
+        tracker.insert_run_id.assert_called_once_with(run_id)
+        client.job_runs.get_job_id_from_run_id.assert_called_once_with(run_id)
+        client.workflow_permissions.patch.assert_called_once_with(
+            "job_id", [{"user_name": "user", "permission_level": "CAN_RUN"}]
+        )
+        client.job_runs.poll_for_completion.assert_called_once_with(run_id)
+        tracker.remove_run_id.assert_called_once_with(run_id)
+
+    def test_submit__with_acls_permission_error(
+        self, submitter, compiled_code, client, tracker, run_id
+    ):
+        job_config = Mock()
+        job_config.job_spec = {
+            "access_control_list": [{"user_name": "user", "permission_level": "CAN_RUN"}]
+        }
+        job_config.additional_job_config = {}
+        submitter.config_compiler.compile.return_value = job_config
+
+        client.job_runs.get_job_id_from_run_id.side_effect = Exception("Error getting job_id")
+
+        with pytest.raises(DbtRuntimeError):
+            submitter.submit(compiled_code)
+
+        tracker.insert_run_id.assert_called_once_with(run_id)
+        client.job_runs.get_job_id_from_run_id.assert_called_once_with(run_id)
+        client.workflow_permissions.put.assert_not_called()
+        tracker.remove_run_id.assert_called_once_with(run_id)
 
     def test_submit__poll_fails__cleans_up(self, submitter, compiled_code, client, tracker, run_id):
+        job_config = Mock()
+        job_config.job_spec = {}
+        job_config.additional_job_config = {}
+        submitter.config_compiler.compile.return_value = job_config
+
         client.job_runs.poll_for_completion.side_effect = Exception("error")
         with pytest.raises(Exception):
             submitter.submit(compiled_code)
@@ -142,7 +194,7 @@ class TestPythonNotebookWorkflowSubmitter:
         submitter.uploader.upload.return_value = "upload_path"
         submitter.config_compiler.compile.return_value = ({}, "existing_job_id")
         submitter.workflow_creater.create_or_update.return_value = "existing_job_id"
-        submitter.permission_builder.build_permissions.return_value = []
+        submitter.permission_builder.build_job_permissions.return_value = []
         submitter.api_client.workflows.run.return_value = "run_id"
         submitter.submit(compiled_code)
         submitter.tracker.insert_run_id.assert_called_once_with("run_id")
@@ -153,7 +205,7 @@ class TestPythonNotebookWorkflowSubmitter:
         submitter.uploader.upload.return_value = "upload_path"
         submitter.config_compiler.compile.return_value = ({}, "existing_job_id")
         submitter.workflow_creater.create_or_update.return_value = "existing_job_id"
-        submitter.permission_builder.build_permissions.return_value = []
+        submitter.permission_builder.build_job_permissions.return_value = []
         submitter.api_client.workflows.run.return_value = "run_id"
         submitter.api_client.job_runs.poll_for_completion.side_effect = Exception("error")
         with pytest.raises(Exception):
@@ -162,14 +214,14 @@ class TestPythonNotebookWorkflowSubmitter:
 
     def test_create__golden_path(self, client, tracker):
         parsed_model = Mock()
+
         parsed_model.config.python_job_config.grants = {}
+        parsed_model.config.access_control_list = []
         parsed_model.config.python_job_config.additional_task_settings = {}
         parsed_model.config.python_job_config.dict.return_value = {}
-        parsed_model.config.access_control_list = []
         submitter = PythonNotebookWorkflowSubmitter.create(client, tracker, parsed_model)
         assert submitter.api_client == client
         assert submitter.tracker == tracker
         assert isinstance(submitter.uploader, PythonNotebookUploader)
         assert isinstance(submitter.config_compiler, PythonWorkflowConfigCompiler)
-        assert isinstance(submitter.permission_builder, PythonPermissionBuilder)
         assert isinstance(submitter.workflow_creater, PythonWorkflowCreator)